+++ packages/block-editor/src/components/global-styles/hooks.js
@@ -219,6 +219,7 @@ export function useColorsPerOrigin( settings ) {
 					'Theme',
 					'Indicates this palette comes from the theme.'
 				),
+				slug: 'theme',
 				colors: themeColors,
 			} );
 		}
@@ -232,6 +233,7 @@ export function useColorsPerOrigin( settings ) {
 					'Default',
 					'Indicates this palette comes from WordPress.'
 				),
+				slug: 'default',
 				colors: defaultColors,
 			} );
 		}
@@ -241,6 +243,7 @@ export function useColorsPerOrigin( settings ) {
 					'Custom',
 					'Indicates this palette is created by the user.'
 				),
+				slug: 'custom',
 				colors: customColors,
 			} );
 		}