+++ packages/components/src/private-apis.ts
@@ -14,6 +14,8 @@ import {
 	ValidatedToggleGroupControl,
 } from './validated-form-controls';
 import { ValidatedFormTokenField } from './validated-form-controls/components/form-token-field';
+import { colorEditingKey } from './color-palette/private-keys';
+import { isColorEditingEnabled } from './color-palette/utils';
 import ContentEditableControl from './content-editable-control';
 
 export const privateApis = {};
@@ -32,4 +34,6 @@ lock( privateApis, {
 	ValidatedToggleControl,
 	ValidatedToggleGroupControl,
 	ValidatedFormTokenField,
+	colorEditingKey,
+	isColorEditingEnabled,
 } );