+++ packages/components/CHANGELOG.md
@@ -51,6 +51,7 @@
 -   Update `@ariakit/react` to 0.4.35 and `@ariakit/test` to 0.7.2 ([#80765](https://github.com/WordPress/gutenberg/pull/80765)).
 -   Update `@floating-ui/react-dom` to 2.1.9 ([#80761](https://github.com/WordPress/gutenberg/pull/80761)).
 -   `ItemGroup`: Migrate styles from Emotion to an SCSS Module. ([#80797](https://github.com/WordPress/gutenberg/pull/80797)).
+-   `ColorPalette`: Add a private API for inline palette color editing. ([#79266](https://github.com/WordPress/gutenberg/pull/79266)).
 
 ## 38.0.0 (2026-07-29)